# Conflicts:
#	.env.example
#	middleware.ts
#	package-lock.json
#	package.json
#	playwright.config.ts

+++ components/invitation/PendingInvitations.tsx
@@ -1,17 +1,26 @@
-import { Card, Error, LetterAvatar, Loading } from '@/components/shared';
+import { Error, LetterAvatar, Loading } from '@/components/shared';
 import { defaultHeaders } from '@/lib/common';
-import { Invitation, Team } from '@prisma/client';
+import { Team } from '@prisma/client';
 import useInvitations from 'hooks/useInvitations';
 import { useTranslation } from 'next-i18next';
-import React from 'react';
-import { Button } from 'react-daisyui';
+import React, { useState } from 'react';
 import toast from 'react-hot-toast';
 import type { ApiResponse } from 'types';
+import ConfirmationDialog from '../shared/ConfirmationDialog';
+import { TeamInvitation } from 'models/invitation';
+import { Table } from '@/components/shared/table/Table';
 
 const PendingInvitations = ({ team }: { team: Team }) => {
-  const { isLoading, isError, invitations, mutateInvitation } = useInvitations(
-    team.slug
-  );
+  const [selectedInvitation, setSelectedInvitation] =
+    useState<TeamInvitation | null>(null);
+
+  const [confirmationDialogVisible, setConfirmationDialogVisible] =
+    useState(false);
+
+  const { isLoading, isError, invitations, mutateInvitation } = useInvitations({
+    slug: team.slug,
+    sentViaEmail: true,
+  });
 
   const { t } = useTranslation('common');
 
@@ -23,13 +32,18 @@ const PendingInvitations = ({ team }: { team: Team }) => {
     return <Error message={isError.message} />;
   }
 
-  const deleteInvitation = async (invitation: Invitation) => {
+  const deleteInvitation = async (invitation: TeamInvitation | null) => {
+    if (!invitation) return;
+
     const sp = new URLSearchParams({ id: invitation.id });
 
-    const response = await fetch(`/api/teams/${team.slug}/invitations?${sp.toString()}`, {
-      method: 'DELETE',
-      headers: defaultHeaders,
-    });
+    const response = await fetch(
+      `/api/teams/${team.slug}/invitations?${sp.toString()}`,
+      {
+        method: 'DELETE',
+        headers: defaultHeaders,
+      }
+    );
 
     const json = (await response.json()) as ApiResponse<unknown>;
 
@@ -47,61 +61,61 @@ const PendingInvitations = ({ team }: { team: Team }) => {
   }
 
   return (
-    <Card heading="Invitations Sent">
-      <Card.Body>
-        <table className="w-full text-left text-sm text-gray-500 dark:text-gray-400">
-          <thead className="bg-gray-50 text-xs uppercase text-gray-700 dark:bg-gray-700 dark:text-gray-400">
-            <tr>
-              <th scope="col" className="px-6 py-3" colSpan={2}>
-                {t('email')}
-              </th>
-              <th scope="col" className="px-6 py-3">
-                {t('role')}
-              </th>
-              <th scope="col" className="px-6 py-3">
-                {t('created-at')}
-              </th>
-              <th scope="col" className="px-6 py-3">
-                {t('action')}
-              </th>
-            </tr>
-          </thead>
-          <tbody>
-            {invitations.map((invitation) => {
-              return (
-                <tr
-                  key={invitation.token}
-                  className="border-b bg-white hover:bg-gray-50 dark:border-gray-700 dark:bg-gray-800 dark:hover:bg-gray-600"
-                >
-                  <td className="px-6 py-3" colSpan={2}>
-                    <div className="flex items-center justify-start space-x-2">
-                      <LetterAvatar name={invitation.email} />
-                      <span>{invitation.email}</span>
-                    </div>
-                  </td>
-                  <td className="px-6 py-3">{invitation.role}</td>
-                  <td className="px-6 py-3">
-                    {new Date(invitation.createdAt).toDateString()}
-                  </td>
-                  <td className="px-6 py-3">
-                    <Button
-                      size="xs"
-                      color="error"
-                      variant="outline"
-                      onClick={() => {
-                        deleteInvitation(invitation);
-                      }}
-                    >
-                      {t('remove')}
-                    </Button>
-                  </td>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      </Card.Body>
-    </Card>
+    <div className="space-y-3">
+      <div className="space-y-3">
+        <h2 className="text-xl font-medium leading-none tracking-tight">
+          {t('pending-invitations')}
+        </h2>
+        <p className="text-sm text-gray-500 dark:text-gray-400">
+          {t('description-invitations')}
+        </p>
+      </div>
+
+      <Table
+        cols={[t('email'), t('role'), t('expires-at'), t('actions')]}
+        body={invitations.map((invitation) => {
+          return {
+            id: invitation.id,
+            cells: [
+              {
+                wrap: true,
+                element: invitation.email ? (
+                  <div className="flex items-center justify-start space-x-2">
+                    <LetterAvatar name={invitation.email} />
+                    <span>{invitation.email}</span>
+                  </div>
+                ) : undefined,
+              },
+              { text: invitation.role },
+              { wrap: true, text: new Date(invitation.expires).toDateString() },
+              {
+                buttons: [
+                  {
+                    color: 'error',
+                    text: t('remove'),
+                    onClick: () => {
+                      setSelectedInvitation(invitation);
+                      setConfirmationDialogVisible(true);
+                    },
+                  },
+                ],
+              },
+            ],
+          };
+        })}
+      ></Table>
+
+      <ConfirmationDialog
+        visible={confirmationDialogVisible}
+        onCancel={() => setConfirmationDialogVisible(false)}
+        onConfirm={() => deleteInvitation(selectedInvitation)}
+        title={t('confirm-delete-member-invitation')}
+      >
+        {t('delete-member-invitation-warning', {
+          email: selectedInvitation?.email,
+        })}
+      </ConfirmationDialog>
+    </div>
   );
 };
